feature: decoupling ImageCacheService via ImageLoader interface Introduce abstraction layer to support multiple image providers (Notes and Satellites).
