+++ src/main/java/com/viking/field_passport_generator/http/LoadedResource.java
@@ -1,0 +1,3 @@
+package com.viking.field_passport_generator.http;
+
+public record LoadedResource(byte[] data, String extension) {}
